Release package building and code structure refactor
